Moved string function into string utils.
