+++ src/core/ecs/entity/system/text_rendering_entity_system.h
@@ -5,10 +5,7 @@
 #include "../../component/components/transform2D_component.h"
 #include "../../component/components/text_label_component.h"
 #include "../../../math/space_handler.h"
-
-struct TextLines {
-    std::vector<std::string> lines = {};
-};
+#include "../../../utils/string_util.h"
 
 class TextRenderingEntitySystem : public EntitySystem {
   private:
@@ -17,28 +14,6 @@ class TextRenderingEntitySystem : public EntitySystem {
     ComponentManager *componentManager = nullptr;
     CameraManager *cameraManager = nullptr;
     std::unordered_map<Entity, TextLines> textLines = {};
-
-  protected:
-    virtual TextLines ConvertNewText(const TextLabelComponent& textLabelComponent) {
-        Logger *logger = Logger::GetInstance();
-        std::vector<std::string> lines = {};
-        if (textLabelComponent.wordWrap && textLabelComponent.text.length() > textLabelComponent.maxCharactersPerLine) {
-            unsigned int lineCount = textLabelComponent.text.length() / textLabelComponent.maxCharactersPerLine;
-            lineCount += textLabelComponent.text.length() % textLabelComponent.maxCharactersPerLine > 0 ? 1 : 0;
-//            logger->Debug(std::string("line count = " + std::to_string(lineCount)));
-            int endPos = 0;
-            for (unsigned int i = 0; i < lineCount; i++) {
-                endPos = i < lineCount - 1 ? textLabelComponent.maxCharactersPerLine : textLabelComponent.text.length() - (i * textLabelComponent.maxCharactersPerLine);
-//                logger->Debug(std::string("end pos = " + std::to_string(endPos)));
-                std::string lineText = textLabelComponent.text.substr(i * textLabelComponent.maxCharactersPerLine, endPos);
-//                logger->Debug("line text = " + lineText);
-                lines.emplace_back(lineText);
-            }
-        } else {
-            lines.emplace_back(textLabelComponent.text);
-        }
-        return TextLines{ lines };
-    }
 
   public:
     TextRenderingEntitySystem() {
@@ -67,7 +42,7 @@ class TextRenderingEntitySystem : public EntitySystem {
 
     void UpdateEntityText(Entity entity, const TextLabelComponent& textLabelComponent) {
         textLines.erase(entity);
-        TextLines entityTextLines = ConvertNewText(textLabelComponent);
+        TextLines entityTextLines = StringUtil::ConvertNewText(textLabelComponent.text, textLabelComponent.wordWrap, textLabelComponent.maxCharactersPerLine);
         textLines.emplace(entity, entityTextLines);
     }
